@@ -31,9 +31,11 @@
         }
     ],
     "require": {
-        "php": ">=5.4"
+        "php": ">=7.4",
+        "friendsofphp/php-cs-fixer": "^2.16"
     },
     "require-dev": {
+        "phpunit/phpunit": "^8.5"
     },
     "autoload": {
         "psr-4": {
